Fixed mathematics of "Go X% way to ..."
The previous version was not mathematically correct.

Previously, when you entered the settings for "Cardi B", set the slider to 50% and gave "Justin Bieber" as input, you would get another result compared to first entering "Justin Bieber", setting to 50% and then "Cardi B", thus indicating that there was a problem. I fixed it using two simple parentheses. 

If you want to predict the baby of two people using this tool, it should be now be correct instead of leaning towards the second name entered.